urukul: work around windows numpy int peculiarity
"OverflowError: Python int too large to convert to C long" otherwise

opticlock#74

+++ artiq/coredevice/urukul.py
@@ -1,7 +1,7 @@
 from artiq.language.core import kernel, delay, portable, at_mu, now_mu
 from artiq.language.units import us, ms
 
-from numpy import int32
+from numpy import int32, int64
 
 from artiq.coredevice import spi2 as spi
 
@@ -175,7 +175,7 @@ class CPLD:
         self.cfg_reg = urukul_cfg(rf_sw=rf_sw, led=0, profile=0,
                                   io_update=0, mask_nu=0, clk_sel=clk_sel,
                                   sync_sel=sync_sel, rst=0, io_rst=0)
-        self.att_reg = int32(att)
+        self.att_reg = int32(int64(att))
         self.sync_div = sync_div
 
     @kernel
